% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTrial.R
\name{simulateTrial}
\alias{simulateTrial}
\title{Simulate a trial with an ordinal outcome.}
\usage{
simulateTrial(oddsRatio, nSamples)
}
\arguments{
\item{oddsRatio}{A scalar that defines the proportional odds ratio between the reference and intervention.}

\item{nSamples}{Size of each trial arm.}
}
\description{
\code{simulateTrial} simulates a trial.
}
\examples{
simulate(oddsRatio = 1.5, nSamples = 100)
}
